[core] Fix package update
This broke with the latest updates to the
package update workflow.

The root cause was that for the affected
functions pkg-desc could be either a package
name or a package description but the function
was expecting a description.

The fix was to have the function test against the
type of the parameter and in the case of a package
name obtain the package description manually.